requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
gunicorn==23.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
redis==7.1.0
python-multipart==0.0.21
google-auth>=2.36.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
pytest-asyncio>=0.25.0
httpx>=0.28.0
requests==2.32.3
groq>=0.4.0
pytest==9.0.2
boto3==1.42.31
alembic==1.18.1
google-genai==1.59.0
pydub==0.25.1
faster-whisper==1.2.1
# TTS: Chatterbox-Turbo (English). PyPI chatterbox-tts does NOT include tts_turbo; install from GitHub to get ChatterboxTurboTTS.
# https://github.com/resemble-ai/chatterbox (src/chatterbox/tts_turbo.py exists only in source)
# git+https://github.com/resemble-ai/chatterbox.git
# Watermarker required by ChatterboxTTS; in some envs PerthImplicitWatermarker is None, we patch to DummyWatermarker
resemble-perth
# IndicF5 TTS for Indic languages (hi, ml, ta). Install with: pip install git+https://github.com/ai4bharat/IndicF5.git
# vocos  # dependency of IndicF5
# pip install peft
# chatterbox-tts